+++ cMake/FreeCAD_Helpers/PrintFinalReport.cmake
@@ -95,7 +95,7 @@ macro(PrintFinalReport)
     value(BUILD_DRAFT)
     value(BUILD_DRAWING)
     value(BUILD_FEM)
-    value(BUILD_WEB)
+    value(BUILD_GUI)
     value(BUILD_HELP)
     value(BUILD_IDF)
     value(BUILD_IMPORT)
