--- src/Gui/PreferencePackTemplates/Start_Colors.cfg
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="no" ?>
-<FCParameters>
-  <FCParamGroup Name="Root">
-    <FCParamGroup Name="BaseApp">
-      <FCParamGroup Name="Preferences">
-        <FCParamGroup Name="Mod">
-          <FCParamGroup Name="Start">
-            <FCUInt Name="BackgroundColor1" Value="1414812927"/>
-            <FCUInt Name="BackgroundTextColor" Value="4294967295"/>
-            <FCUInt Name="PageColor" Value="690563583"/>
-            <FCUInt Name="PageTextColor" Value="3789677055"/>
-            <FCUInt Name="BoxColor" Value="1296911871"/>
-            <FCUInt Name="LinkColor" Value="3099197439"/>
-            <FCUInt Name="BackgroundColor2" Value="2141107711"/>
-          </FCParamGroup>
-        </FCParamGroup>
-      </FCParamGroup>
-    </FCParamGroup>
-  </FCParamGroup>
-</FCParameters>
