+++ src/Gui/PreferencePacks/Classic/Classic.cfg
@@ -35,15 +35,6 @@
             <FCUInt Name="DefaultBBoxSelectionColor" Value="3372220415"/>
             <FCUInt Name="DefaultBBoxNormalColor" Value="4294967295"/>
           </FCParamGroup>
-          <FCParamGroup Name="Start">
-            <FCUInt Name="BackgroundColor1" Value="1331197183"/>
-            <FCUInt Name="BackgroundTextColor" Value="1600086015"/>
-            <FCUInt Name="PageColor" Value="4294967295"/>
-            <FCUInt Name="PageTextColor" Value="255"/>
-            <FCUInt Name="BoxColor" Value="3722305023"/>
-            <FCUInt Name="LinkColor" Value="65535"/>
-            <FCUInt Name="BackgroundColor2" Value="2141107711"/>
-          </FCParamGroup>
           <FCParamGroup Name="Spreadsheet">
             <FCText Name="AliasedCellBackgroundColor">#feff9e</FCText>
             <FCText Name="TextColor">#000000</FCText>
