+++ src/Gui/PreferencePacks/Dark contrast/Dark contrast.cfg
@@ -23,15 +23,6 @@
           <FCParamGroup Name="Spreadsheet">
             <FCText Name="AliasedCellBackgroundColor">#9b4de6</FCText>
           </FCParamGroup>
-          <FCParamGroup Name="Start">
-            <FCUInt Name="BackgroundColor1" Value="1010580735"/>
-            <FCUInt Name="BackgroundTextColor" Value="4294703103"/>
-            <FCUInt Name="PageColor" Value="858993663"/>
-            <FCUInt Name="PageTextColor" Value="3570717951"/>
-            <FCUInt Name="BoxColor" Value="1583243007"/>
-            <FCUInt Name="LinkColor" Value="1437270015"/>
-            <FCUInt Name="BackgroundColor2" Value="2141107711"/>
-          </FCParamGroup>
         </FCParamGroup>
         <FCParamGroup Name="OutputWindow">
           <FCUInt Name="colorText" Value="3570717951"/>
