+++ src/Gui/PreferencePacks/Light modern/Light modern.cfg
@@ -4,15 +4,6 @@
     <FCParamGroup Name="BaseApp">
       <FCParamGroup Name="Preferences">
         <FCParamGroup Name="Mod">
-          <FCParamGroup Name="Start">
-            <FCUInt Name="BackgroundColor1" Value="4143380223"/>
-            <FCUInt Name="BackgroundTextColor" Value="4294703103"/>
-            <FCUInt Name="PageColor" Value="4294967295"/>
-            <FCUInt Name="PageTextColor" Value="255"/>
-            <FCUInt Name="BoxColor" Value="3974950143"/>
-            <FCUInt Name="LinkColor" Value="1437270015"/>
-            <FCUInt Name="BackgroundColor2" Value="2141107711"/>
-          </FCParamGroup>
           <FCParamGroup Name="Spreadsheet">
             <FCText Name="AliasedCellBackgroundColor">#feff9e</FCText>
           </FCParamGroup>
